+++ lca-team6/src/components/nav/NavBarShow.jsx
